# backend/public/openapi.yaml
openapi: 3.1.0
info:
  title: Cultural Site API
  version: 1.0.0
  description: API for managing cultural sites, proposals, reviews, and user data.
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Review:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB ObjectId for the review.
        culturalSite:
          type: string
          description: ObjectId of the CulturalSite being reviewed.
        user:
          type: string
          description: ObjectId of the User who wrote the review.
        rating:
          type: number
          format: int32
          minimum: 1
          maximum: 5
          description: Rating given by the user (1 to 5).
        comment:
          type: string
          maxLength: 500
          description: Comment provided by the user.
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the review was created.
        active:
          type: boolean
          default: true
          description: Field for soft deletion. (not implemented)
      required:
        - culturalSite
        - user
        - rating
...
paths:
...
  /cultural-sites/{culturalSiteId}/reviews/{reviewId}:
    get:
      summary: Get a specific review for a cultural site
      tags:
        - Reviews
      operationId: getReviewByIdForCulturalSite
      parameters:
        - in: path
          name: culturalSiteId
          schema:
            type: string
          required: true
          description: Unique ID of the cultural site.
        - in: path
          name: reviewId
          schema:
            type: string
          required: true
          description: Unique ID of the review.
      responses:
        200:
          description: Details of the review.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        404:
          description: Cultural site or review not found.
        500:
          description: Internal server error.
    patch:
      summary: Update a specific review
      tags:
        - Reviews
      operationId: updateReviewById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: culturalSiteId
          schema:
            type: string
          required: true
          description: Unique ID of the cultural site.
        - in: path
          name: reviewId
          schema:
            type: string
          required: true
          description: Unique ID of the review to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: number
                  format: int32
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
                  maxLength: 500
            examples:
              updateReview:
                value:
                  comment: 'An updated comment.'
      responses:
        200:
          description: Review updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        401:
          description: Unauthorized.
        403:
          description: Forbidden, only the review's owner can update it.
        404:
          description: Review not found.
        500:
          description: Internal server error.
    delete:
      summary: Delete a specific review
      tags:
        - Reviews
      operationId: deleteReviewById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: culturalSiteId
          schema:
            type: string
          required: true
          description: Unique ID of the cultural site.
        - in: path
          name: reviewId
          schema:
            type: string
          required: true
          description: Unique ID of the review to delete.
      responses:
        204:
          description: Review successfully deleted.
        401:
          description: Unauthorized.
        403:
          description: Forbidden, only the review's owner or an admin can delete it.
        404:
          description: Review not found.
        500:
          description: Internal server error.
...